Handling of user id, GeneratedPlaylist
This commit does 2 things:
First, change the handling of user ids. As the current user can not
easily be changed (only through logging out of the Spotify account in
browser and looging in with a different user), it doesn't make sense to
make the user configurable for any operation that involves user
authorization. This change removes the argument for those functions and
uses the user_id from the config instead.

Second, implement the GeneratedPlaylist type, a class that manages a
generated playlist built from defined criteria. GeneratedPlaylist tales
as an argument a function, iterates all songs in the Library and uses
the function to determine whether this song os supposed to be added. It
also creates the playlist via the spotify client.